// serophin/mobile/src/navigation/types.ts
export type RootStackParamList = {
  Home: undefined;
  Meditation: undefined;
  MeditationSession: undefined;
  Breathing: undefined;
  BreathingSession: undefined;
  Sleep: undefined;
  SleepSession: undefined;
  Course: undefined;
  CourseLesson: { level: number; lesson: number };
  Settings: undefined;
};